If applied, this commit will add to the users' collection database once, if the user signs in through Google.
