+++ src/components/Login.js
@@ -1,51 +1,14 @@
 import React, { Component } from 'react';
-import { login, resetPassword, googleLogin, saveUser } from '../helpers/auth';
+import { login, resetPassword, googleLogin } from '../helpers/auth';
 
 import RaisedButton from 'material-ui/RaisedButton';
 import TextField from 'material-ui/TextField';
-import { firebaseAuth } from '../config/constants';
 
 function setErrorMsg(error) {
   return {
     loginMessage: error
   };
 }
-
-function onGoogleLoginReload() {
-  // Result from Redirect auth flow.
-  firebaseAuth().getRedirectResult()
-  .then(result => {
-    // if (result.credential) {
-    //   // This gives you a Google Access Token. You can use it to access the Google API.
-    //   const token = result.credential.accessToken;
-    //   document.getElementById('quickstart-oauthtoken').textContent = token;
-    // } else {
-    //   document.getElementById('quickstart-oauthtoken').textContent = 'null';
-    // }
-    // The signed-in user info.
-    const user = result.user;
-    saveUser(user);
-  })
-  .catch(error => {
-    const errorCode = error.code;
-    const errorMessage = error.message;
-    // The email of the user's account used.
-    const email = error.email;
-    console.log('email: ', email);
-    // The firebase.auth.AuthCredential type that was used.
-    var credential = error.credential;
-    console.log('credential: ', credential);
-    console.log(`${errorCode}: ${errorMessage}`);
-    if (errorCode === 'auth/account-exists-with-different-credential') {
-      alert('You have already signed up with a different auth provider for that email.');
-      // If you are using multiple auth providers on your app you should handle linking
-      // the user's accounts here.
-    } else {
-      console.error(error);
-    }
-  });
-}
-
 
 export default class Login extends Component {
   constructor(props) {
@@ -55,11 +18,6 @@ export default class Login extends Component {
       password: '',
       loginMessage: null
     };
-  }
-
-  componentDidMount() {
-    console.log('hit');
-    onGoogleLoginReload();
   }
 
   handleSubmit = e => {
