+++ src/components/protected/Dashboard.js
@@ -1,6 +1,49 @@
 import React, { Component } from 'react';
+import { saveUser } from '../../helpers/auth';
+import { db, firebaseAuth } from '../../config/constants';
+
+function onGoogleLoginReload() {
+  // Result from Redirect auth flow.
+  return firebaseAuth().getRedirectResult()
+    .then(result => {
+      const user = result.user;
+      db.collection('users')
+      .where('email', '==', user.email)
+      .get()
+      .then(snapshot => {
+        console.log('snapshot: ', snapshot);
+        if (snapshot.docs.length === 0) {
+          saveUser(user);
+        } else {
+          snapshot.forEach(doc => {
+            console.log('Document data:', doc.data());
+          })
+        }
+      })
+    })
+    .catch(error => {
+      const errorCode = error.code;
+      const errorMessage = error.message;
+      // The email of the user's account used.
+      const email = error.email;
+      // The firebase.auth.AuthCredential type that was used.
+      var credential = error.credential;
+      if (errorCode === 'auth/account-exists-with-different-credential') {
+        alert('You have already signed up with a different auth provider for that email.');
+        // If you are using multiple auth providers on your app you should handle linking
+        // the user's accounts here.
+      } else {
+        console.error(error);
+      }
+    });
+}
 
 export default class Dashboard extends Component {
+
+  componentDidMount() {
+    onGoogleLoginReload();
+  }
+
   render() {
     return (
       <div>
